Finish render functionality
